Add the remaining Unpacker test cases
